1. Fix generate_tag function 2. Fix localization 3. add confirm on delete-all-mappings-button

+++ js/admin.js
@@ -1510,6 +1510,9 @@ jQuery(document).ready(function($) {
 		
 		clearMessages();
 		
+		if(!confirm(wpcf7_pdf_forms.__Confirm_Delete_All_Mappings))
+			return;
+		
 		deleteAllMappings();
 		
 		return false;
